week-16/pyproject.toml
[tool.poetry]
name = "ssafy-custom-news-backend"
version = "0.1.0"
description = ""
authors = ["jiwoochris <travelandi01@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1.2"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
gunicorn = "^23.0.0"
dj-rest-auth = "^6.0.0"
django-allauth = "^65.1.0"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.32.3"
django-health-check = "^3.18.3"
django-cors-headers = "^4.5.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
django-stubs = "^5.1.0"
black = "^24.10.0"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
python_version = "3.10"